+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+
+* Added restart function.
+* Updated retry function.
+
+
 ## 0.0.2
 
 * Minor bug fixes.